% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{detect_comments_column}
\alias{detect_comments_column}
\title{Detect the free-text comments column}
\usage{
detect_comments_column(df)
}
\arguments{
\item{df}{A cleaned data frame.}
}
\value{
The name of the comments column.
}
\description{
Detect the free-text comments column
}
