% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_clean_form.R
\name{read_clean_form}
\alias{read_clean_form}
\title{Read and clean ICM training quality assessment data}
\usage{
read_clean_form(path)
}
\arguments{
\item{path}{Path to the Excel file exported from M365 Forms.}
}
\value{
A cleaned tibble with derived variables.
}
\description{
Read and clean ICM training quality assessment data
}
