% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_hospital_report.R
\name{render_hospital_report}
\alias{render_hospital_report}
\title{Render a report for a single hospital}
\usage{
render_hospital_report(
  hospital_name,
  df_clean,
  regional_summary,
  output_file,
  template = system.file("rmd", "hospital_report_template.Rmd", package =
    "icmBenchmarkR")
)
}
\arguments{
\item{hospital_name}{Name of the hospital.}

\item{df_clean}{Cleaned dataset.}

\item{regional_summary}{Output of summarise_region().}

\item{output_file}{Full path to the PDF to be created.}

\item{template}{Path to the Rmd template.}
}
\description{
Render a report for a single hospital
}
